Add dotfiles example package
This package has a setup-dotfiles binary, which when executed links all
the files of the dotfiles/home directory into the users $HOME via
symlinks. The symlinks point to the paths within the profile where the
dotfiles package was installed, which means they automatically update if
newer versions of this package are installed.

However:
- If a new file was added to dotfiles/home, the command needs to be
  reran by the users
- If a file was deleted from dotfiles/home, the file needs to be
  manually removed by the users
- If a file in dotfiles/home already exists in the users $HOME, nothing
  is done for this file. So users need to manually remove preexisting files if
  they wish to use this approach